Placing common intersection operations in their own file to remove duplicated definitions
