+++ include/xdg/geometry/face_common.h
@@ -2,6 +2,7 @@
 #define XDG_GEOMETRY_TRIANGLE_INTERSECT_H
 
 #include <array>
+#include "xdg/shared_enums.h"
 #include "xdg/vec3da.h"
 
 namespace xdg {
